backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.85.0",
    "fastapi>=0.135.1",
    "google-genai",
    "httpx>=0.28.1",
    "openai>=2.28.0",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "router>=0.1",
    "uvicorn>=0.42.0",
]
